add using secret to login in and modify app apis
