+++ ssolib/models/app/app.go
@@ -15,7 +15,6 @@ import (
 	"github.com/laincloud/sso/ssolib/models"
 	"github.com/laincloud/sso/ssolib/models/group"
 	"github.com/laincloud/sso/ssolib/models/iuser"
-	"github.com/laincloud/sso/ssolib/models/role"
 )
 
 var (
@@ -199,47 +198,3 @@ func AppNameExist(ctx *models.Context, appName string) (bool, error) {
 	}
 }
 
-func DeleteApp(ctx *models.Context, id int) ( error) {
-	tx := ctx.DB.MustBegin()
-	roles := []role.Role{}
-	err1 := ctx.DB.Select(&roles, "SELECT id FROM role WHERE app_id=?", id)
-	if err1 != nil {
-		return err1
-	}
-	_, err2 := tx.Exec("DELETE FROM role WHERE app_id=?", id)
-	if err2 != nil {
-		return err2
-	}
-	_, err3 := tx.Exec("DELETE FROM resource WHERE app_id=?", id)
-	if err3 != nil {
-		return err3
-	}
-	for _,r := range roles {
-		if role.IsLeafRole(ctx, r.Id) {
-			_, err := tx.Exec("DELETE FROM role_resource WHERE role_id=?", id)
-			if err != nil {
-				return err
-			}
-		}
-	}
-	_, err4 := tx.Exec("DELETE FROM app WHERE id=?", id)
-	if err4 != nil {
-		return err4
-	}
-	for _,r := range roles {
-		g, err := group.GetGroup(ctx, r.Id)
-		if err != nil {
-			log.Error(err)
-			panic(err)
-		}
-		err = group.DeleteGroup(ctx, g)
-		if err != nil {
-			panic(err)
-		}
-	}
-	if err5 := tx.Commit(); err5 != nil {
-		log.Debug(err5)
-		return err5
-	}
-	return nil
-}
