+++ ssolib/models/role/app.go
@@ -83,3 +83,48 @@ func GetAppAdminRole(ctx *models.Context, appId int) (*Role, error) {
 	}
 	return GetRole(ctx, app.AdminGroupId)
 }
+
+func DeleteApp(ctx *models.Context, id int) ( error) {
+	tx := ctx.DB.MustBegin()
+	roles := []Role{}
+	err1 := ctx.DB.Select(&roles, "SELECT id FROM role WHERE app_id=?", id)
+	if err1 != nil {
+		return err1
+	}
+	_, err2 := tx.Exec("DELETE FROM role WHERE app_id=?", id)
+	if err2 != nil {
+		return err2
+	}
+	_, err3 := tx.Exec("DELETE FROM resource WHERE app_id=?", id)
+	if err3 != nil {
+		return err3
+	}
+	for _,r := range roles {
+		if IsLeafRole(ctx, r.Id) {
+			_, err := tx.Exec("DELETE FROM role_resource WHERE role_id=?", id)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	_, err4 := tx.Exec("DELETE FROM app WHERE id=?", id)
+	if err4 != nil {
+		return err4
+	}
+	for _,r := range roles {
+		g, err := group.GetGroup(ctx, r.Id)
+		if err != nil {
+			log.Error(err)
+			panic(err)
+		}
+		err = group.DeleteGroup(ctx, g)
+		if err != nil {
+			panic(err)
+		}
+	}
+	if err5 := tx.Commit(); err5 != nil {
+		log.Debug(err5)
+		return err5
+	}
+	return nil
+}
